retry when sqlite throws ErrorIO
I suspect this may be due to SQLITE_IOERR_SHORT_READ, but have not
verified.

I was able to reproduce it on Linux after running the test suite in a loop
for 1-3 hours until it failed.

The WAL mode entry change in 3963c5fcf5b568408719663cb51e870704e8634a
may have hidden the problem I was seeing; I have not seen an ErrorIO
since then.